+++ th2_data_services/provider/v5/provider_api/__init__.py
@@ -12,10 +12,5 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from .data import Data
-from .data_source import DataSource
-from .filter import Filter
-import logging
-from logging import NullHandler
-
-logging.getLogger(__name__).addHandler(NullHandler())
+from .grpc import GRPCProvider5API
+from .http import HTTPProvider5API