+++ th2_data_services/source_api.py
@@ -12,10 +12,8 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from .data import Data
-from .data_source import DataSource
-from .filter import Filter
-import logging
-from logging import NullHandler
+from abc import ABC
 
-logging.getLogger(__name__).addHandler(NullHandler())
+
+class ISourceAPI(ABC):
+    """High level interface for Source API."""