+++ release_notes.md
@@ -142,3 +142,17 @@
 4. [TH2-2656] Added new optional parameter to EventsTree (ParentEventsTree) classes.
     - Set "preserve_body=True" to keep event bodies during tree building.
     - Otherwise events bodies will be omitted.
+
+# v0.6.1
+
+## BugFixes
+
+1. [TH2-2858] Resolved HTTP error 414 on a client side.
+2. [TH2-2851] UnicodeDecodeError is raised during events extracting if not utf-8 character in the byte stream. 
+User will get original byte now.
+
+# v0.6.2
+
+## BugFixes
+
+1. [TH2-2922] Fixed infinite loop in events tree.